# apps/user-web-api/documentation/openapi-docs.yaml
openapi: 3.0.0
info:
  title: Facebluk User Command API
  version: 1.0.0
  description: These are the APIs exposed by our backend SDK. To be consumed by the frontend only.
servers:
  - url: http://localhost:3000/api
    description: API
tags:
  - name: Posts
  - name: Friend Requests
paths:
  /posts/create:
    post:
      operationId: createPost
      summary: Create a new user post
      description: |
        With this route you can create a new Facebluk post
        for your authenticated user
      tags:
        - Posts
      requestBody:
        description: Payload
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
                - userId
              properties:
                description:
                  type: string
                  maxLength: 500
                userId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: The user post was created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - postId
                properties:
                  postId:
                    type: string
                    format: uuid
        '422':
          $ref: '#/components/responses/BusinessRuleError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /friend-requests/send:
    post:
      operationId: sendFriendRequest
      summary: Sends a new friend request
      description: |
        With this route you can send a friend request
        from a user to another one
      tags:
        - Friend Requests
      requestBody:
        description: Payload
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fromUserId
                - toUserId
              properties:
                fromUserId:
                  type: string
                  format: uuid
                toUserId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: The friend request was sent successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - friendRequestId
                properties:
                  friendRequestId:
                    type: string
                    format: uuid
        '422':
          $ref: '#/components/responses/BusinessRuleError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /friend-requests/accept:
    post:
      operationId: acceptFriendRequest
      summary: Accepts a pending friend request
      description: |
        With this route you can accept a pending friend 
        request
      tags:
        - Friend Requests
      requestBody:
        description: Payload
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - friendRequestId
              properties:
                userId:
                  type: string
                  format: uuid
                friendRequestId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: The friend request was accepted successfully
        '422':
          $ref: '#/components/responses/BusinessRuleError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /friend-requests/cancel:
    post:
      operationId: cancelFriendRequest
      summary: Cancels a pending friend request
      description: |
        With this route you can cancel a pending friend 
        request
      tags:
        - Friend Requests
      requestBody:
        description: Payload
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - friendRequestId
              properties:
                userId:
                  type: string
                  format: uuid
                friendRequestId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: The friend request was cancelled successfully
        '422':
          $ref: '#/components/responses/BusinessRuleError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /friend-requests/reject:
    post:
      operationId: rejectFriendRequest
      summary: Rejects a pending friend request
      description: |
        With this route you can reject a pending friend 
        request
      tags:
        - Friend Requests
      requestBody:
        description: Payload
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - friendRequestId
              properties:
                userId:
                  type: string
                  format: uuid
                friendRequestId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: The friend request was cancelled successfully
        '422':
          $ref: '#/components/responses/BusinessRuleError'
        '500':
          $ref: '#/components/responses/UnexpectedError'

components:
  responses:
    BusinessRuleError:
      description: BusinessRuleError
      content:
        application/json:
          schema:
            type: object
            required:
              - requestId
              - message
            properties:
              requestId:
                type: string
                format: uuid
              message:
                type: string
    UnexpectedError:
      description: UnexpectedError
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string